Biggest refactoring in history of Atlas

* commit '4e5e8010d82e51235daf2feded66748720702205': (74 commits)
  Fix atlas_fctest_functionspace
  ATLAS-68: Cleanup Fortran interface
  ATLAS-68: Cleanup Fortran interface
  ATLAS-68: Cleanup Fortran interface
  ATLAS-68: Cleanup Fortran interface
  ATLAS-68: Cleanup Fortran interface
  Fix BuildCellCentres
  Compilation on Cray
  Optionally  compile deprecated::FunctionSpace
  Deprecate old FunctionSpace, and rename next::FunctionSpace to FunctionSpace
  Deprecate old FunctionSpace in Fortran interface
  ATLAS-68: Cleanup Fortran interface
  ATLAS-56: Convert atlas_fctest_nabla_EdgeBasedFiniteVolume to new design
  ATLAS-56: Expose cells and edges in atlas_Mesh Fortran interface
  ATLAS-56: atlas_mesh_Elements and atlas_mesh_ElementType Fortran interfaces
  ATLAS-68: Cleanup Fortran interface
  ATLAS-56: Start porting atlas_mesh_HybridElements to Fortran
  DEBUG macros work without mpi
  ATLAS-56: Fortran - atlas_MultiBlockConnectivity, atlas_BlockConnectivity
  ATLAS-56: Export missing value to fortran atlas_Connectivity
  ...

+++ src/atlas/FieldSet.cc
@@ -103,7 +103,7 @@ int    atlas__FieldSet__has_field     (FieldSet* This, char* name)
   return 0;
 }
 
-int    atlas__FieldSet__size          (FieldSet* This)
+size_t atlas__FieldSet__size          (FieldSet* This)
 {
   ATLAS_ERROR_HANDLING(
     ASSERT(This != NULL);
@@ -121,7 +121,7 @@ Field* atlas__FieldSet__field_by_name (FieldSet* This, char* name)
   return NULL;
 }
 
-Field* atlas__FieldSet__field_by_idx  (FieldSet* This, int idx)
+Field* atlas__FieldSet__field_by_idx  (FieldSet* This, size_t idx)
 {
   ATLAS_ERROR_HANDLING(
     ASSERT(This != NULL);
